Update run unit tests paths
