+++ apps/vs-code-designer/src/app/utils/unitTests.ts
@@ -16,6 +16,14 @@ export interface UnitTestResult {
   duration?: number;
 }
 
+/**
+ * Saves the unit test definition for a workflow.
+ * @param {string} projectPath The path of the project.
+ * @param {string} workflowName The name of the workflow.
+ * @param {string} unitTestName The name of the unit test.
+ * @param {any} unitTestDefinition The unit test definition.
+ * @returns A Promise that resolves when the unit test definition is saved.
+ */
 export const saveUnitTestDefinition = async (
   projectPath: string,
   workflowName: string,
